+++ 2-back-end-javaspring/src/main/java/vn/gt/__back_end_javaspring/service/impl/ShareServiceImpl.java
@@ -133,7 +133,10 @@ public class ShareServiceImpl implements ShareService {
             throw new UserNotFoundException("You are not allowed to delete this share");
         }
 
-        earningEventService.deleteShareEvent(share.getId());
+        if(reviewerService.isReviewerByUserId(userId)){
+            earningEventService.deleteShareEvent(share.getId());
+
+        }
 
         share.setIsDeleted(true);
         shareRepository.save(share);
